Create Bank Customer Classification
This Dataset  contains a list of  bank customers who have a credit card with the bank. It suggests whether a credit card customer is performing good or bad for a bank. It contains 20 features including demographic, socio-economic and transactional variables for 5050 customers who have an active credit card with the bank.
                 Here I examined each variables, summarized the finding and interpreted them. I detected the  outliers in 22 variables and treated them. I treated the  missing values in overall 22 variables. I Developed a logistic regression and Random Forest model to estimate the relationship between 'class' and the other variables.